requirements.txt
# VideoEditor GUI Agent Dependencies
gradio_client>=0.8.0
tkinterdnd2>=0.3.0
requests>=2.28.0
Pillow>=9.0.0
jsonschema>=4.0.0

# Video Editing Agent Dependencies
# Note: uuid, datetime, pathlib, dataclasses, typing are built-in Python modules

# Optional: For enhanced functionality
python-dotenv>=0.19.0
opencv-python>=4.5.0  # For video metadata extraction in fallback mode
google-genai>=0.2.0  # For Gemini AI image analysis and prompt enhancement

# Additional dependencies for robust error handling
urllib3>=1.26.0

# Minimal API server for testing
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Web UI support
python-multipart>=0.0.9
Jinja2>=3.1.0
